src/main/java/com/azs/beefygainz/exercise/service/ExerciseServiceImpl.java
package com.azs.beefygainz.exercise.service;

import com.azs.beefygainz.exercise.exception.NoSuchExerciseException;
import com.azs.beefygainz.exercise.model.Exercise;
import com.azs.beefygainz.exercise.repository.ExerciseRepository;
import org.springframework.beans.factory.annotation.Autowired;

import java.util.ArrayList;
import java.util.List;

public class ExerciseServiceImpl implements ExerciseService {

    private final ExerciseRepository exerciseRepository;

    @Autowired
    public ExerciseServiceImpl(ExerciseRepository exerciseRepository) {
        this.exerciseRepository = exerciseRepository;
    }

    @Override
    public List<Exercise> findAllByUserId(long userId) {
        List<Exercise> exercises = new ArrayList<>();

        exerciseRepository.findAllByUserId(userId).forEach(exercises::add);

        return exercises;
    }

    @Override
    public Exercise save(Exercise exercise) {
        return exerciseRepository.save(exercise);
    }

    @Override
    public Exercise findById(long id) {
        return exerciseRepository.findById(id).orElseThrow(() -> new NoSuchExerciseException(id));
    }
}
